+++ lib/rivet/utils.rb
@@ -8,43 +8,62 @@ module Rivet
 
     # This returns the merged definition given a group
 
-    def self.get_definition(config, group, directory)
+    def self.get_definition(group, directory)
       defaults   = consume_defaults(directory)
-      group_def  = load_definition(group, directory)
-      config_def = load_definition(group, directory, config) if config
+      group_def  = load_definition(group, File.join(directory, 'groups'))
 
-      # Deep merge group_def into the defaults or use group_def
-      # NOTE: This will replace any arrays in default_def with group_def arrays
-      definition = if defaults && group_def
-        defaults.deep_merge(group_def)
-      else
-        group_def
-      end
+      # A successful rivet run requires a group_def, if none, bail out...
+      if group_def
+        definition =  if group_def['include']
+                        # group_def['include'] can be an array of strings or a string (if containing only one common_def)
+                        # If group_def includes common_def(s), merge them (in order)
+                        #   1) defaults (if any) deep merge with first common_def
+                        #   2) If common_defs, deep merge merged_result with subsequent common_defs
+                        #   3) After last common_def deep merge, deep merge result with group_def (takes precendence)
+                        #      with :concat_array option
+                        # If no common_defs
+                        #   1) Deep merge group_def into defaults (if any)
+                        result =  if group_def['include'].respond_to? :empty
+                                    group_def['include'].inject({}) do |h, (filename_path, index)|
+                                      filename_path << '.yml' unless filename_path[-4..-1] == '.yml'
+                                      common_def  = load_definition('common', directory, filename_path)
+                                      h = if index == 0
+                                            defaults ? deep_merge(defaults, common_def) : common_def
+                                          else
+                                            deep_merge(r, common_def)
+                                          end
+                                    end
+                                  else
+                                    filename_path = group_def['include']
+                                    filename_path << '.yml' unless filename_path[-4..-1] == '.yml'
+                                    common_def  = load_definition('common', directory, filename_path)
 
-      # Now Deep merge the group's config_def into the definition using a block:
-      # NOTE: This will concat any arrays in definition with config_def arrays and
-      #       will remove duplicates (uniq), but user really should't be
-      #       creating dups...
-      definition = definition.deep_merge(config_def) do |k, tv, v|
-        tv.is_a?(Array) ? tv.concat(v).uniq : v
-      end if definition && config_def
+                                    defaults ? deep_merge(defaults, common_def) : common_def
+                                  end
+                        deep_merge(result, group_def, :concat_arrays)
+                      else
+                        defaults ? deep_merge(defaults, group_def) : group_def
+                      end
 
-      # Handle 'run_list' edge-case
-      # NOTE: * It doesn't remove dups again, so don't create any (user responsiblity
-      #         and yes it's possible by creating two runlist entries with differnt
-      #         ordering positions.
-      #       * config_def takes precedence on ordering position
-      definition['bootstrap']['run_list'] = begin
-        definition['bootstrap']['run_list'].inject([]) do |h, v|
-          if v.respond_to? :each
-            h.insert(v[1], v[0])
-          else
-            h << v
+        # Handle 'run_list' edge-case
+        # NOTE: * It doesn't remove dups again, so don't create any (user responsiblity
+        #         and yes it's possible by creating two runlist entries with differnt
+        #         ordering positions.
+        #       * config_def takes precedence on ordering position
+        definition['bootstrap']['run_list'] = begin
+          definition['bootstrap']['run_list'].inject([]) do |h, v|
+            if v.respond_to? :each
+              h.insert(v[1], v[0])
+            else
+              h << v
+            end
           end
-        end
-      end if definition && definition['bootstrap'] && definition['bootstrap']['run_list']
+        end if definition && definition['bootstrap'] && definition['bootstrap']['run_list']
 
-      definition ? definition : false
+        definition
+      else
+        false
+      end
     end
 
     # Gobbles up the defaults file from YML, returns the hash or false if empty
@@ -83,5 +102,25 @@ module Rivet
       end
     end
 
+    # This helper method returns the result of a deep_merge that either
+    # replaces arrays or concats arrays
+
+    def self.deep_merge(hash, other_hash, type=:replace_arrays)
+      if type == :replace_arrays
+        # Deep merge other_hash into the hash
+        # NOTE: This will replace any arrays in hash with other_hash arrays
+        hash.deep_merge(other_hash)
+      elsif type == :concat_arrays
+        # Now Deep merge the other_hash into the hash using a block:
+        # NOTE: This will concat any arrays in hash with other_hash arrays and
+        #       will remove duplicates (uniq), but user really should't be
+        #       creating dups...
+        hash.deep_merge(other_hash) do |k, tv, v|
+          tv.is_a?(Array) ? tv.concat(v).uniq : v
+        end
+      else
+        Rivet::Utils.die("InvalidArgument for Utils.deep_merge: 'type' => #{type}")
+      end
+    end
   end
 end